Move config, dqrs packages to internal path
Move "internal" packages to an actual internal path.
