+++ internal/dqrs/constants.go
