+++ internal/dqrs/convert.go
