+++ internal/dqrs/doc.go
