+++ internal/dqrs/queryresponse.go
