+++ internal/dqrs/record.go
