+++ internal/dqrs/summary.go
